fix(tui): splitter hover/drag highlights with per-splitter state and clippy clean
- Add SplitterHoverState to AppState with 6 hover + 6 drag booleans (per splitter).
- Implement update_splitter_hover() hit-testing for all visible splitters on MouseMove.
- Preserve drag flags across hover recomputes; set/clear them on MouseDown/MouseUp.
- Wire hover/drag through render chain (app, explorer, discover, sql_tab, history).
- Fix swapped param order in splitter::view::render causing hover to light wrong line.
- Fix splitter_at_with_detail to include EditorResults hit when detail is visible.
- Add unit tests: hit-test coverage, v/h splitter non-overlap, hover color rendering.
- Apply clippy auto-fixes (collapsible_if, question_mark, matches!, etc.).
- Add #[allow(clippy::too_many_arguments)] on 4 render functions (param count is
  semantically justified). Allow field_reassign_with_default for test code only.
- All 601 tests pass; full workspace clippy: zero warnings.

+++ crates/dbm-tui2/src/features/explorer/instances/view.rs
@@ -349,7 +349,7 @@ mod tests {
                 line.push_str(buf[(x, y)].symbol());
             }
             if line.contains("a") && !line.contains("Instances") {
-                found_y = Some(y as u16);
+                found_y = Some(y);
                 break;
             }
         }
